backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
qdrant-client==1.7.0
pydantic==2.5.3
pydantic[email]
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
asyncpg==0.29.0
httpx==0.26.0
alembic==1.13.1
psycopg2-binary
PyJWT[crypto]==2.9.0
openai-chatkit
